Move Error typings to common
