+++ x-pack/plugins/infra/server/lib/adapters/metrics/adapter_types.ts
@@ -111,7 +111,3 @@ export type InfraMetricModelCreator = (
   indexPattern: string | string[],
   interval: string
 ) => InfraMetricModel;
-
-export enum InfraMetricsErrorCodes {
-  invalid_node = 'METRICS_INVALID_NODE',
-}
